App 100% Bn gaya hai AI kehra hai abhi, sab chiz done hai AI ke taraf se

+++ lib/core/services/storage_service.dart
@@ -183,8 +183,8 @@ final storageStatsProvider = FutureProvider<StorageStats>((ref) async {
     svc.getDownloadSizeBytes(),
   ]);
   return StorageStats(
-    cacheSizeBytes:      results[0] as int,
-    downloadedSongCount: results[1] as int,
-    downloadSizeBytes:   results[2] as int,
+    cacheSizeBytes:      results[0],
+    downloadedSongCount: results[1],
+    downloadSizeBytes:   results[2],
   );
 });
